Travel/requirements.txt
flask
flask-cors
requests
pandas
python-dotenv
scikit-learn
